requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
numpy==1.20.1
Pillow==8.1.0
psycopg2==2.8.6
Pygments==2.7.3
PySocks==1.7.1
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
whitenoise==5.2.0
